[MIN-153] Remove setThreshold.
