+++ src/actions/economicData.ts
@@ -1,9 +1,6 @@
 import { web3FromAddress } from '@polkadot/extension-dapp';
 import { Dispatch } from '../util/types';
 import {
-	SET_THRESHOLD_REQUEST_ERROR,
-	SET_THRESHOLD_REQUEST_SUCCESS,
-	SET_THRESHOLD_REQUEST_START,
 	RESET_ADMIN_REQUESTS,
 	SET_LIQUIDATIONS_MAX_ATTEMPTS_ERROR,
 	SET_LIQUIDATIONS_MAX_ATTEMPTS_START,
@@ -35,7 +32,7 @@ import {
 } from './types';
 import API from '../services';
 import { UNDERLYING_ASSETS_TYPES } from '../util/constants';
-import { txCallback, convertInputToPercent } from '../util';
+import { txCallback } from '../util';
 
 export function setLiquidationMaxAttempts(
 	account: string,
@@ -80,44 +77,6 @@ export function setLiquidationMaxAttempts(
 export const resetAdminRequests = () => {
 	return {
 		type: RESET_ADMIN_REQUESTS,
-	};
-};
-
-export const setThreshold = (
-	account: string,
-	keyring: any,
-	poolId: string,
-	newAmount: string
-) => {
-	return async (dispatch: Dispatch) => {
-		const callBack = txCallback(
-			[SET_THRESHOLD_REQUEST_SUCCESS, SET_THRESHOLD_REQUEST_ERROR],
-			dispatch
-		);
-
-		try {
-			dispatch({ type: SET_THRESHOLD_REQUEST_START });
-			const currentUser = keyring.getPair(account);
-			const convertThreshold = convertInputToPercent(newAmount);
-
-			if (currentUser.isLocked) {
-				const injector = await web3FromAddress(account);
-				await API.tx.sudo
-					.sudo(API.tx.riskManager.setThreshold(poolId, convertThreshold))
-					// @ts-ignore
-					.signAndSend(account, { signer: injector.signer }, callBack);
-			} else {
-				await API.tx.sudo
-					.sudo(API.tx.riskManager.setThreshold(poolId, convertThreshold))
-					// @ts-ignore
-					.signAndSend(currentUser, callBack);
-			}
-		} catch (err) {
-			dispatch({
-				type: SET_THRESHOLD_REQUEST_ERROR,
-				payload: err.toString(),
-			});
-		}
 	};
 };
